+++ app/models/user.rb
@@ -12,6 +12,7 @@ class User < ActiveRecord::Base
 
   # Model Relationships
   has_many :projects, -> { where deleted: false }
+  has_many :categories, -> { where deleted: false }
 
   # User Methods